requirements.txt
Django
psycopg2
djangorestframework
requests
django-environ
drf-nested-routers
django-filter